Added tests for the text_preparation_pipeline.py, merged regex tests into test_normalizers.
